fix: add scheduledAt filter and column logic
