+++ src/core_modules/capture-core/components/ListView/types/listView.types.js
@@ -26,6 +26,7 @@ export type FilterOnly = {
     multiValueFilter?: boolean,
     tooltipContent?: string,
     disabled?: boolean,
+    transformRecordsFilter: (rawFilter: any) => Object
 };
 
 export type AdditionalFilter = {
