+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/Setup/hooks/index.js
@@ -1,0 +1,10 @@
+// @flow
+export { useDefaultColumnConfig } from './useDefaultColumnConfig';
+export { useStaticTemplates } from './useStaticTemplates';
+export { useFiltersOnly } from './useFiltersOnly';
+export { useProgramStageFilters } from './useProgramStageFilters';
+export {
+    useInjectDataFetchingMetaToLoadList,
+    useInjectDataFetchingMetaToUpdateList,
+} from './useInjectDataFetchingMeta';
+export { useFiltersToKeep } from './useFiltersToKeep';
