+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/Setup/hooks/useStaticTemplates.js
@@ -1,0 +1,62 @@
+// @flow
+import { useMemo } from 'react';
+
+export const useStaticTemplates = () =>
+    useMemo(
+        () => [
+            {
+                id: 'default',
+                isDefault: true,
+                name: 'default',
+                access: {
+                    update: false,
+                    delete: false,
+                    write: false,
+                    manage: false,
+                },
+            },
+            {
+                id: 'active',
+                name: 'Active enrollments',
+                order: 1,
+                access: {
+                    update: false,
+                    delete: false,
+                    write: false,
+                    manage: false,
+                },
+                criteria: {
+                    programStatus: 'ACTIVE',
+                },
+            },
+            {
+                id: 'complete',
+                name: 'Completed enrollments',
+                order: 2,
+                access: {
+                    update: false,
+                    delete: false,
+                    write: false,
+                    manage: false,
+                },
+                criteria: {
+                    programStatus: 'COMPLETED',
+                },
+            },
+            {
+                id: 'cancelled',
+                name: 'Cancelled enrollments',
+                order: 3,
+                access: {
+                    update: false,
+                    delete: false,
+                    write: false,
+                    manage: false,
+                },
+                criteria: {
+                    programStatus: 'CANCELLED',
+                },
+            },
+        ],
+        [],
+    );
