+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/helpers/eventFilters/additionalFiltersConvertor.js
@@ -2,12 +2,14 @@
 export const ADDITIONAL_FILTERS = {
     programStage: 'programStage',
     occurredAt: 'eventOccurredAt',
+    scheduledAt: 'scheduledAt',
     status: 'status',
 };
 
 export const ADDITIONAL_FILTERS_LABELS = {
     programStage: 'Program stage',
     occurredAt: 'Report date',
+    scheduledAt: 'Scheduled date',
     status: 'Event status',
 };
